+++ src/PublicServiceRegistry.Api.Backoffice/LabelType/LabelTypeController.cs
@@ -13,6 +13,7 @@ namespace PublicServiceRegistry.Api.Backoffice.LabelType
     using Security;
     using Swashbuckle.AspNetCore.Filters;
     using LabelType = PublicServiceRegistry.LabelType;
+    using ProblemDetails = Be.Vlaanderen.Basisregisters.BasicApiProblem.ProblemDetails;
 
     [ApiVersion("1.0")]
     [AdvertiseApiVersions("1.0")]
@@ -29,7 +30,7 @@ namespace PublicServiceRegistry.Api.Backoffice.LabelType
         /// <response code="500">Als er een interne fout is opgetreden.</response>
         [HttpGet]
         [ProducesResponseType(typeof(List<LabelTypeListResponse>), StatusCodes.Status200OK)]
-        [ProducesResponseType(typeof(BasicApiProblem), StatusCodes.Status500InternalServerError)]
+        [ProducesResponseType(typeof(ProblemDetails), StatusCodes.Status500InternalServerError)]
         [SwaggerResponseExample(StatusCodes.Status200OK, typeof(LabelTypeListResponseExamples), jsonConverter: typeof(StringEnumConverter))]
         [SwaggerResponseExample(StatusCodes.Status500InternalServerError, typeof(InternalServerErrorResponseExamples), jsonConverter: typeof(StringEnumConverter))]
         [AllowAnonymous]