+++ src/PublicServiceRegistry.Projections.Backoffice/Migrations/20190524194412_AddIpdcCode.cs
@@ -1,0 +1,24 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+namespace PublicServiceRegistry.Projections.Backoffice.Migrations
+{
+    public partial class AddIpdcCode : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<string>(
+                name: "IpdcCode",
+                schema: "PublicServiceRegistryBackoffice",
+                table: "PublicServiceList",
+                nullable: true);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "IpdcCode",
+                schema: "PublicServiceRegistryBackoffice",
+                table: "PublicServiceList");
+        }
+    }
+}