+++ src/PublicServiceRegistry/PublicService/Exceptions/InsufficientRights.cs
@@ -7,4 +7,13 @@ namespace PublicServiceRegistry.PublicService.Exceptions
         public InsufficientRights()
             : base(ExceptionMessage) { }
     }
+
+    public class IpdcCodeWasAlreadySet: PublicServiceRegistryException
+    {
+        private const string ExceptionMessage = "Op deze dienstverlening is reeds een ipdc code toegewezen. " +
+                                                "Gelieve eerst de vorige ipdc code te verwijderen vooraleer een nieuwe toe te kennen.";
+
+        public IpdcCodeWasAlreadySet()
+            : base(ExceptionMessage) { }
+    }
 }