+++ src/PublicServiceRegistry/PublicService/PublicServiceCommandHandlerModule.cs
@@ -119,6 +119,19 @@ namespace PublicServiceRegistry.PublicService
 
                     publicService.Remove(message.Command.ReasonForRemoval);
                 });
+
+            For<SetIpdcCode>()
+                .AddSqlStreamStore(getStreamStore, getUnitOfWork, eventMapping, eventSerializer)
+                .RequiresRole(PublicServiceRegistryClaims.AdminRole)
+                .Handle(async (message, ct) =>
+                {
+                    var publicServices = getPublicServices();
+
+                    var publicServiceId = message.Command.PublicServiceId;
+                    var publicService = await publicServices.GetAsync(publicServiceId, ct);
+
+                    publicService.SetIpdcCode(message.Command.IpdcCode);
+                });
         }
     }
 }